Code Refractor + Optimized the Comunication Interface + made some referral endpoint functional (#61)
* Tests

* communication: ArrayConvertible Testing

* communication: ArrayConvertible Testing

* communication: .

* communication: .

* Code Refraktor

* Code Refraktor

* Code Refactor

* communication: .

* communication: .

* communication: .

* Code Refactor

* communication: Added Groups

* Code Refactor

* Code Refactor

* Code Refactor

* Code Refactor

* Update RouteCollector.php

* Updateded SQL

* communication: mysql

* communication: Bugfix

* communication: Bugfix

* Update UserRegisterValidator.php

* Code Restruckturing

* MySQL Changed

* MySQL Changed

* Code Refractor

* Code Refractor

* Code Refractor

* Code Refractor

* Code Refractor + Made some Endpoint functional

+++ src/Repository/ReferralRepository/MySqlReferralRepository.php
@@ -5,77 +5,81 @@ declare(strict_types=1);
 namespace Nebalus\Webapi\Repository\ReferralRepository;
 
 use Nebalus\Webapi\Exception\ApiException;
+use Nebalus\Webapi\Repository\AbstractRepository;
 use Nebalus\Webapi\Value\ID;
 use Nebalus\Webapi\Value\Referral\Referral;
 use Nebalus\Webapi\Value\Referral\ReferralCode;
+use Nebalus\Webapi\Value\Referral\ReferralId;
+use Nebalus\Webapi\Value\Referral\ReferralName;
 use Nebalus\Webapi\Value\Referral\ReferralPointer;
 use Nebalus\Webapi\Value\Referral\Referrals;
+use Nebalus\Webapi\Value\User\UserId;
 use PDO;
 
-class MySqlReferralRepository
+class MySqlReferralRepository extends AbstractRepository
 {
     public function __construct(
-        private readonly PDO $pdo,
-        private readonly RedisReferralCachingRepository $redis
+        private PDO $pdo,
+        private RedisReferralCachingRepository $redis
     ) {
+        parent::__construct($pdo);
     }
 
-    public function insertReferral(ID $userId, ReferralCode $code, ReferralPointer $pointer, bool $disabled = true): bool
+    public function insertReferral(UserId $ownerUserId, ReferralCode $code, ReferralPointer $pointer, ReferralName $name, bool $disabled = true): bool
     {
-        $sql = "INSERT INTO referrals(user_id, code, pointer, disabled) VALUES (:user_id, :code, :pointer, :disabled)";
+        $sql = "INSERT INTO referrals(owner_user_id, code, pointer, name, disabled) VALUES (:owner_user_id, :code, :pointer, :name, :disabled)";
 
         $stmt = $this->pdo->prepare($sql);
-        $stmt->bindValue(':user_id', $userId->asInt());
+        $stmt->bindValue(':owner_user_id', $ownerUserId->asInt());
         $stmt->bindValue(':code', $code->asString());
         $stmt->bindValue(':pointer', $pointer->asString());
+        $stmt->bindValue(':name', $name->asString());
         $stmt->bindValue(':disabled', $disabled);
         return $stmt->execute();
     }
 
-    public function insertReferralClickEntry(ID $referralId): bool
+    public function insertReferralClickEntry(ReferralId $referralId): bool
     {
-        $sql = "INSERT INTO referral_clicks(referral_id) VALUES (:referral_id)";
+        $sql = "INSERT INTO referral_click_metric(referral_id) VALUES (:referral_id)";
 
         $stmt = $this->pdo->prepare($sql);
         $stmt->bindValue(':referral_id', $referralId->asInt());
         return $stmt->execute();
     }
 
-    public function deleteReferralById(ID $referralId): bool
+    public function deleteReferralByCodeAndOwnerId(ReferralCode $code, UserId $ownerUserId): bool
     {
-        $sql = "DELETE FROM referrals WHERE referral_id = :referral_id";
-
-        $stmt = $this->pdo->prepare($sql);
-        $stmt->bindValue(':referral_id', $referralId->asInt());
-        $stmt->execute();
-
-        return $stmt->rowCount() === 1;
-    }
-
-    public function deleteReferralByCode(ReferralCode $code): bool
-    {
-        $sql = "DELETE FROM referrals WHERE code = :code";
+        $sql = "DELETE FROM referrals WHERE code = :code AND owner_user_id = :owner_user_id";
 
         $stmt = $this->pdo->prepare($sql);
         $stmt->bindValue(':code', $code->asString());
+        $stmt->bindValue(':owner_user_id', $ownerUserId->asInt());
         $stmt->execute();
 
         return $stmt->rowCount() === 1;
     }
 
-    public function updateReferral()
+    public function updateReferral(Referral $referral, UserId $ownerUserId): bool
     {
+        $sql = "UPDATE `referrals` SET `pointer`=:pointer,`name`=:name,`disabled`=:disabled WHERE `owner_user_id`=:owner_user_id AND `code`=:code";
+
+//        $stmt = $this->pdo->prepare($sql);
+//        $stmt->bindValue(':pointer', $referral->asString());
+//        $stmt->bindValue(':owner_user_id', $ownerUserId->asInt());
+//        $stmt->execute();
+
+        return false;
     }
 
     /**
      * @throws ApiException
      */
-    public function getReferralsFromUserId(ID $userId): Referrals
+    public function getReferralsFromOwnerId(UserId $ownerUserId): Referrals
     {
-        $sql = "SELECT * FROM referrals WHERE user_id = :user_id";
+        $sql = "SELECT * FROM referrals WHERE owner_user_id = :owner_user_id";
 
         $stmt = $this->pdo->prepare($sql);
-        $stmt->bindValue(':user_id', $userId->asInt());
+        $stmt->bindValue(':owner_user_id', $ownerUserId->asInt());
         $stmt->execute();
 
         $data = [];
@@ -90,12 +94,12 @@ class MySqlReferralRepository
     /**
      * @throws ApiException
      */
-    public function findReferralById(ID $id): ?Referral
+    public function findReferralByCode(ReferralCode $code): ?Referral
     {
-        $sql = "SELECT * FROM referrals WHERE referral_id = :referral_id";
+        $sql = "SELECT * FROM referrals WHERE code = :code";
 
         $stmt = $this->pdo->prepare($sql);
-        $stmt->bindValue(':referral_id', $id->asInt());
+        $stmt->bindValue(':code', $code->asString());
         $stmt->execute();
 
         $data = $stmt->fetch();
@@ -110,12 +114,13 @@ class MySqlReferralRepository
     /**
      * @throws ApiException
      */
-    public function findReferralByCode(ReferralCode $code): ?Referral
+    public function findReferralByCodeAndOwnerId(ReferralCode $code, UserId $ownerUserId): ?Referral
     {
-        $sql = "SELECT * FROM referrals WHERE BINARY code = :code";
+        $sql = "SELECT * FROM referrals WHERE code = :code AND owner_user_id = :owner_user_id";
 
         $stmt = $this->pdo->prepare($sql);
         $stmt->bindValue(':code', $code->asString());
+        $stmt->bindValue(':owner_user_id', $ownerUserId->asInt());
         $stmt->execute();
 
         $data = $stmt->fetch();
